+++ sdk/communication/azure-communication-common/src/main/java/com/azure/android/communication/common/CommunicationIdentifier.java
@@ -28,9 +28,7 @@ public abstract class CommunicationIdentifier {
      * @throws IllegalArgumentException raw id is null or empty.
      */
     public static CommunicationIdentifier fromRawId(String rawId) {
-        if (rawId == null || rawId.trim().length() == 0) {
-            throw new IllegalArgumentException("The parameter [rawId] cannot be null to empty.");
-        }
+        StringUtils.validateNotNullOrEmpty(rawId, "rawId");
 
         if (rawId.startsWith(PHONE_NUMBER_PREFIX)) {
             return new PhoneNumberIdentifier(rawId.substring(PHONE_NUMBER_PREFIX.length()));
@@ -60,11 +58,12 @@ public abstract class CommunicationIdentifier {
                 return new MicrosoftTeamsAppIdentifier(suffix, CommunicationCloudEnvironment.DOD);
             case TEAMS_APP_GCCH_CLOUD_PREFIX:
                 return new MicrosoftTeamsAppIdentifier(suffix, CommunicationCloudEnvironment.GCCH);
-            case ACS_USER_PREFIX:
             case SPOOL_USER_PREFIX:
+                return new CommunicationUserIdentifier(rawId);
+            case ACS_USER_PREFIX:
             case ACS_USER_DOD_CLOUD_PREFIX:
             case ACS_USER_GCCH_CLOUD_PREFIX:
-                return new CommunicationUserIdentifier(rawId);
+                return tryCreateTeamsExtensionUserOrCommunicationUser(prefix, suffix, rawId);
             default:
                 return new UnknownIdentifier(rawId);
         }
@@ -109,5 +108,39 @@ public abstract class CommunicationIdentifier {
     @Override
     public int hashCode() {
         return getRawId().hashCode();
+    }
+
+    private static CommunicationIdentifier tryCreateTeamsExtensionUserOrCommunicationUser(String prefix, String suffix,
+                                                                                          String rawId) {
+        String[] segments = suffix.split("_");
+        if (segments.length != 3) {
+            return new CommunicationUserIdentifier(rawId);
+        }
+
+        String resourceId = segments[0];
+        String tenantId = segments[1];
+        String userId = segments[2];
+        CommunicationCloudEnvironment cloud = determineCloudEnvironment(prefix);
+
+        return new TeamsExtensionUserIdentifier(userId, tenantId, resourceId).setCloudEnvironment(cloud);
+    }
+
+    /**
+     * Determine the cloud based on identifier prefix.
+     * @param cloudPrefix .
+     * @return CommunicationCloudEnvironment.
+     * @throws IllegalArgumentException thrown if CommunicationCloudEnvironment cannot be initialized.
+     */
+    static CommunicationCloudEnvironment determineCloudEnvironment(String cloudPrefix) {
+        switch (cloudPrefix) {
+            case ACS_USER_DOD_CLOUD_PREFIX:
+                return CommunicationCloudEnvironment.DOD;
+            case ACS_USER_GCCH_CLOUD_PREFIX:
+                return CommunicationCloudEnvironment.GCCH;
+            case ACS_USER_PREFIX:
+                return CommunicationCloudEnvironment.PUBLIC;
+            default:
+                throw  new IllegalArgumentException("Cannot initialize CommunicationCloudEnvironment.");
+        }
     }
 }